Add files via upload
Icon added, as well as content js. I think it should work like: the popup.js button listens for a click, then it will pass an array of links to the content.js for datastructure-fying and then the popup.html will display them? I am learning how browser extensions work and this could be totally wrong!